Move local backend under library/providers/ (issue #27)
Introduces src/library/providers/ to separate concrete backend
implementations from shared infrastructure (db, media, bundle).
LocalLibrary moves to providers/local.rs; factory.rs is the only
caller and its import path is updated. Immich backend will slot in
as providers/immich.rs when implemented.
