+++ src/library.rs
@@ -6,8 +6,8 @@ pub mod event;
 pub mod factory;
 pub mod import;
 pub mod importer;
-pub mod local;
 pub mod media;
+pub mod providers;
 pub mod storage;
 
 pub use error::LibraryError;
