+++ src/library/providers/local.rs
@@ -5,14 +5,14 @@ use async_trait::async_trait;
 use tokio::runtime::Handle;
 use tracing::{debug, info, instrument};
 
-use super::bundle::Bundle;
-use super::db::Database;
-use super::error::LibraryError;
-use super::event::LibraryEvent;
-use super::import::LibraryImport;
-use super::importer::ImportJob;
-use super::media::{LibraryMedia, MediaId, MediaRecord};
-use super::storage::LibraryStorage;
+use crate::library::bundle::Bundle;
+use crate::library::db::Database;
+use crate::library::error::LibraryError;
+use crate::library::event::LibraryEvent;
+use crate::library::import::LibraryImport;
+use crate::library::importer::ImportJob;
+use crate::library::media::{LibraryMedia, MediaId, MediaRecord};
+use crate::library::storage::LibraryStorage;
 
 /// Local filesystem backend.
 ///
